defined-49/refactor: расширение правил eslint
