+++ src/pages/Leaderboard/components/LeaderboardTable/LeaderboardTable.scss
@@ -92,7 +92,7 @@
 
     &_type_increase::after {
       content: "";
-      background-image: url("../../img/leaderboard-sort-button.svg");
+      background-image: url("../../../../img/leaderboard-sort-button.svg");
       position: absolute;
       width: 25px;
       height: 25px;
@@ -105,7 +105,7 @@
 
     &_type_decrease::after {
       content: "";
-      background-image: url("../../img/leaderboard-sort-button.svg");
+      background-image: url("../../../../img/leaderboard-sort-button.svg");
       position: absolute;
       width: 25px;
       height: 25px;
