+++ src/pages/Login/Login.tsx
@@ -7,7 +7,7 @@ import Form from '../../components/Form/Form';
 import { FormInput } from '../../components/FormInput/FormInput';
 import Spacer from '../../components/Spacer/Spacer';
 import { useAuth } from '../../hook/useAuth';
-import { PATTERN_VALIDATION } from '../../utils/Const';
+import { PATTERN_VALIDATION } from '../../utils/constants/validation';
 
 const Login = () => {
   const navigate = useNavigate();
